implement milestone item component
